+++ cashu/mint/protocols.py
@@ -1,4 +1,4 @@
-from typing import Dict, Mapping, Protocol
+from typing import Dict, Mapping, Protocol, List
 
 from ..core.base import Method, MintKeyset, Unit
 from ..core.crypto.secp import PublicKey
@@ -10,9 +10,14 @@ from .db.write import DbWriteHelper
 from .events.events import LedgerEventManager
 
 
+class SupportsSeed(Protocol):
+    seed: str
+
 class SupportsKeysets(Protocol):
+    amounts: List[int]
     keyset: MintKeyset
     keysets: Dict[str, MintKeyset]
+    derivation_path: str
 
 
 class SupportsBackends(Protocol):
